[refactoring] New project tree
