+++ src/test/java/com/sistearth/core/routes/IndexGetRouteTest.java
@@ -1,7 +1,8 @@
-package com.sistearth.core.index.routes;
+package com.sistearth.core.routes;
 
 import com.sistearth.api.handler.Answer;
 import com.sistearth.api.handler.Handler;
+import com.sistearth.core.routes.IndexGetRoute;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
