+++ src/test/java/com/sistearth/core/routes/PostsGetRouteTest.java
@@ -1,7 +1,8 @@
-package com.sistearth.core.posts.routes;
+package com.sistearth.core.routes;
 
 import com.sistearth.api.handler.Answer;
 import com.sistearth.api.handler.Handler;
+import com.sistearth.core.routes.PostsGetRoute;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
